+++ app/config.py
@@ -1,5 +1,8 @@
+from fastapi.templating import Jinja2Templates
+
 # general
 DOMAIN = 'Our-Domain'
+templates = Jinja2Templates(directory="templates")
 
 # export
 ICAL_VERSION = '2.0'
