+++ app/dependencies.py
@@ -1,0 +1,9 @@
+from app.database.database import SessionLocal
+
+
+def get_db():
+    db = SessionLocal()
+    try:
+        yield db
+    finally:
+        db.close()
