+++ app/internal/share/share_event.py
@@ -3,9 +3,9 @@ from typing import List, Dict, Union
 from sqlalchemy.orm import Session
 
 from app.database.models import Event, Invitation, UserEvent
-from app.utils.export import event_to_ical
-from app.utils.user import does_user_exist, get_users
-from app.utils.utils import save
+from app.internal.share.export import event_to_ical
+from app.internal.user import does_user_exist, get_users
+from app.internal.utils import save
 
 
 def sort_emails(
