+++ app/main.py
@@ -1,29 +1,17 @@
 import uvicorn
-from fastapi import Depends, FastAPI, Form, Request
-from fastapi.responses import RedirectResponse
+from fastapi import FastAPI, Request
 from fastapi.staticfiles import StaticFiles
-from fastapi.templating import Jinja2Templates
-from sqlalchemy.orm import Session
-from starlette.status import HTTP_302_FOUND
 
-from app.database.database import SessionLocal, engine
+from app.config import templates
+from app.database.database import engine
 from app.database.models import Base
-from app.internal.share_event import accept
-from app.utils.invitation import get_all_invitations, get_invitation_by_id
+from app.routers import invitation
 
 app = FastAPI()
 app.mount("/static", StaticFiles(directory="static"), name="static")
-templates = Jinja2Templates(directory="templates")
+app.include_router(invitation.invitation)
 
 Base.metadata.create_all(bind=engine)
-
-
-def get_db():
-    db = SessionLocal()
-    try:
-        yield db
-    finally:
-        db.close()
 
 
 @app.get("/")
@@ -32,28 +20,6 @@ def home(request: Request):
         "request": request,
         "message": "Hello, World!"
     })
-
-
-@app.get("/invitations")
-def view_invitations(request: Request, db: Session = Depends(get_db)):
-    return templates.TemplateResponse("invitations.html", {
-        "request": request,
-        # recipient_id should be the current user
-        # but because we don't have one yet,
-        # "get_all_invitations" returns all invitations
-        "invitations": get_all_invitations(session=db),
-        "message": "Hello, World!"
-    })
-
-
-@app.post("/invitations")
-async def accept_invitations(
-        invite_id: int = Form(...),
-        db: Session = Depends(get_db)
-):
-    invitation = get_invitation_by_id(invite_id, session=db)
-    accept(invitation, db)
-    return RedirectResponse("/invitations", status_code=HTTP_302_FOUND)
 
 
 @app.get("/profile")
