+++ tests/test_export.py
@@ -1,7 +1,7 @@
 from icalendar import vCalAddress
 
 from app.config import ICAL_VERSION, PRODUCT_ID
-from app.utils.export import (
+from app.internal.share.export import (
     create_ical_calendar, create_ical_event, event_to_ical
 )
