+++ tests/test_invitation.py
@@ -1,4 +1,4 @@
-from app.utils.invitation import get_all_invitations, get_invitation_by_id
+from app.internal.invitation import get_all_invitations, get_invitation_by_id
 
 
 class TestInvitations:
